requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-pptx==1.0.2
python-docx==1.1.0
Pillow==11.3.0
PyMuPDF==1.26.4
requests==2.31.0
python-dotenv
openai
google-generativeai
simplejson

